refactor(api): reorganización completa de llamadas a la API
- Centralización de todos los llamados a la API.
- Creación de axiosPrivate para endpoints autenticados.
- Refactor de CRUD:
  - membresías
  - asignación de   membresías
  - usuarios diarios
  - usuarios mensuales
- Reestructuración de rutas y archivos que consumen las APIs
