+++ gimnasioReact/src/api/action/notifications.api.ts
@@ -1,0 +1,10 @@
+import { Notification } from "../../model/notifications.model"
+import { axiosPrivate } from "../axios.private";
+
+
+export const getMemberNotifications = async (): Promise<Notification[]> => {
+    // const token = localStorage.getItem('token');   
+    const response = await axiosPrivate.get<Notification[]>('/membership-notifications/')
+    return response.data;
+    
+}
